Add testing suite of files
I have covered the central cases of a page with an image, a page without an image, a page that is not a blog post (so shouldn't be copied), a page with a broken image (to see if that fails gracefully), a page with wikilinks, and a page with prettified wikilinks. We could write an explicit testing and verification code.